// src/theme/palette/types.ts
export interface MainColors {
  crimson: string;
  brown: string;
  red: string;
  gray: string;
}

export interface CrimsonColors {
  crimson: string;
  crimsonDarken10: string;
  crimsonDarken20: string;
  crimsonDarken30: string;
  crimsonLighten10: string;
  crimsonLighten20: string;
  crimsonLighten30: string;
  crimsonLighten40: string;
  crimsonLighten50: string;
  crimsonLighten60: string;
  crimsonLighten70: string;
  crimsonLighten80: string;
  crimsonLighten90: string;
  crimsonLighten100: string;
}

export interface GrayColors {
  gray: string;
  grayDarken10: string;
  grayDarken20: string;
  grayDarken30: string;
  grayLighten10: string;
  grayLighten20: string;
  grayLighten30: string;
  grayLighten40: string;
  grayLighten50: string;
  grayLighten60: string;
  grayLighten70: string;
  grayLighten80: string;
  grayLighten90: string;
  grayLighten100: string;
}

export interface BrownColors {
  brown: string;
  brownDarken10: string;
  brownDarken20: string;
  brownDarken30: string;
  brownLighten10: string;
  brownLighten20: string;
  brownLighten30: string;
  brownLighten40: string;
  brownLighten50: string;
  brownLighten60: string;
  brownLighten70: string;
  brownLighten80: string;
  brownLighten90: string;
  brownLighten100: string;
}

export interface RedColors {
  red: string;
  redDarken10: string;
  redDarken20: string;
  redDarken30: string;
  redLighten10: string;
  redLighten20: string;
  redLighten30: string;
  redLighten40: string;
  redLighten50: string;
  redLighten60: string;
  redLighten70: string;
  redLighten80: string;
  redLighten90: string;
  redLighten100: string;
}

export type Palette = CrimsonColors & GrayColors & BrownColors & RedColors;